requirements.txt
certifi==2026.2.25
numpy==2.4.4
pandas==3.0.2
protobuf==7.34.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
ta==0.11.0
tzdata==2026.1
upstox-python-sdk==2.21.0
urllib3==2.6.3
uuid==1.30
websocket-client==1.9.0
Flask==3.0.2
flasgger==0.9.7.1
requests
gunicorn==21.2.0
APScheduler==3.10.4
